% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R
\name{barcode_plotter}
\alias{barcode_plotter}
\title{Barcode plotter}
\usage{
barcode_plotter(Ydat, Xdat, maxscale = rep(0.05, ncol(Xdat)))
}
\arguments{
\item{maxscale}{}
}
\value{

}
\description{
Barcode plotter
}
