% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot method for objects \code{TopSA}}
\usage{
plot(TopSAObj, nvar = "all", ...)
}
\arguments{
\item{TopSAObj}{an object of class \code{TopSA}}

\item{nvar}{it could be a secuence from 1 to the number of variables
indicating which variables should be plotted. It could be the character
'all' for plot all the variables.}

\item{...}{further arguments passed to the \code{plot} function}
}
\value{
A plot of generated with the output of \code{\link{TopSA}} . For each
variable in the model, it creates the plot of the correponding complex, its
symmetric reflection and its symmetric difference.
}
\description{
Plot the Complex, symmetric reflection and symmetric diffrence for and object
of class \code{\ĺink{TopSA}}
}
\examples{

ishigami.fun <- function(X) {
A <- 7
B <- 0.1
sin(X[, 1]) + A * sin(X[, 2])^2 + B * X[, 3]^4 * sin(X[, 1])
}

X <- matrix(runif(3*100, -pi, pi), ncol = 3)
Y <- ishigami.fun(X)

estimation <- sobolnp(Y = Y, X = X, nboot = 5)

plot(estimation)

}
