% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.TopSA}
\title{Plot method for objects \code{TopSA}}
\usage{
print(TopSAObj, ...)

\method{print}{TopSA}(TopSAObj, only.return.table = FALSE, ...)
}
\arguments{
\item{TopSAObj}{an object of class \code{TopSA}}

\item{...}{further arguments passed to the \code{plot} function}

\item{only.return.table}{if \code{TRUE}, returns a data frame with the
estimated values. Otherwise, print the data frame in console. Defaults to
\code{FALSE}}
}
\value{
A plot of generated with the output of \code{\link{TopSA}} . For each
variable in the model, it creates the plot of the correponding complex, its
symmetric reflection and its symmetric difference.
}
\description{
Plot the Complex, symmetric reflection and symmetric diffrence for and object
of class \code{\ĺink{TopSA}}
}
\examples{

ishigami.fun <- function(X) {
A <- 7
B <- 0.1
sin(X[, 1]) + A * sin(X[, 2])^2 + B * X[, 3]^4 * sin(X[, 1])
}

X <- matrix(runif(3*100, -pi, pi), ncol = 3)
Y <- ishigami.fun(X)

estimation <- sobolnp(Y = Y, X = X, nboot = 5)

plot(estimation)


}
