% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopSA.R
\name{TopSA}
\alias{TopSA}
\title{Title}
\usage{
TopSA(Ydat, Xdat, dimension = 3, threshold = 0.05, knearest = 20,
  method = c("delanauy", "VR"), mc.cores = 1)
}
\arguments{
\item{Ydat}{numeric vector of responses in a model.}

\item{Xdat}{numeric matrix or data.frame of covariables.}

\item{dimension}{number of homology groups to estimate. The only value
accepted is \code{dimension = 3}, but in future release this will change.}

\item{threshold}{percent of radius or sizes of triangles to keep to smooth
the homology complex. Defaults to `3`.}

\item{knearest}{number of the nearest neigbors keep to built the homology
complex}

\item{method}{type of method to build the homology complex. Two choices are
accepted: \code{delanauy} o \code{VR} (Vietoris-Rips).}

\item{mc.cores}{number of cores to use in the procedure. Defaults \code{1}.}
}
\value{
Put output
}
\description{
Title
}
\examples{

ishigami.fun <- function(X) {
A <- 7
B <- 0.1
sin(X[, 1]) + A * sin(X[, 2])^2 + B * X[, 3]^4 * sin(X[, 1])
}

X <- matrix(runif(3*100, -pi, pi), ncol = 3)
Y <- ishigami.fun(X)

estimation <- TopSA(Ydat = Y, Xdat = X)

}
